Scenario 2: Add the base structure and steps for scenario 2.
Added two additional steps to the feature file to ensure that the notification is correctly shown and split entering the address and submitting changes (to increase step reusability)

Since last step references the third one, this information has to be stored between the steps somehow. The ScenarioContext is the class dedicated just for storing this information between the steps.